+++ src/Core/Render/Twig/HandlebarsTokenParser.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * @package Picowind
+ * @subpackage Picowind
+ * @since 1.0.0
+ */
+
+namespace Picowind\Core\Render\Twig;
+
+use Twig\Token;
+use Twig\TokenParser\AbstractTokenParser;
+
+/**
+ * Parses {% handlebars %} tags in Twig templates
+ *
+ * Syntax:
+ *   {% handlebars 'template.hbs' %}
+ *   {% handlebars 'template.hbs' with {'var': 'value'} %}
+ *   {% handlebars 'template.hbs' with {'var': 'value'} only %}
+ */
+class HandlebarsTokenParser extends AbstractTokenParser
+{
+    public function parse(Token $token): HandlebarsNode
+    {
+        $lineno = $token->getLine();
+        $stream = $this->parser->getStream();
+
+        $template = $this->parser->parseExpression();
+        $with = null;
+        $only = false;
+
+        if ($stream->nextIf(Token::NAME_TYPE, 'with')) {
+            $with = $this->parser->parseExpression();
+        }
+
+        if ($stream->nextIf(Token::NAME_TYPE, 'only')) {
+            $only = true;
+        }
+
+        $stream->expect(Token::BLOCK_END_TYPE);
+
+        return new HandlebarsNode($template, $with, $only, $lineno, $this->getTag());
+    }
+
+    public function getTag(): string
+    {
+        return 'handlebars';
+    }
+}